# apps/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ВСМ Практика API",
    "version": "1.0.0",
    "description": "Cookie auth. Все времена в UTC. Состояние и начисления хранятся в PostgreSQL."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "vsm_session"
      }
    },
    "responses": {
      "Error": {
        "description": "Ошибка API",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiFailure"
            }
          }
        }
      }
    },
    "schemas": {
      "ApiFailure": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "description": "Необязательные подробности"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "isDemo": {
            "type": "boolean"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "student",
              "author",
              "admin"
            ]
          },
          "brigade": {
            "type": "string"
          },
          "depot": {
            "type": "string"
          },
          "company": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "brigade",
          "depot",
          "company"
        ],
        "additionalProperties": false
      },
      "ModuleFlags": {
        "type": "object",
        "properties": {
          "account": {
            "type": "boolean"
          },
          "play": {
            "type": "boolean"
          },
          "editor": {
            "type": "boolean"
          },
          "notifications": {
            "type": "boolean"
          },
          "progress": {
            "type": "boolean"
          },
          "team": {
            "type": "boolean"
          },
          "leaderboard": {
            "type": "boolean"
          },
          "immersive": {
            "type": "boolean"
          },
          "voice": {
            "type": "boolean"
          },
          "materials": {
            "type": "boolean"
          }
        },
        "required": [
          "account",
          "play",
          "editor",
          "notifications",
          "progress",
          "team",
          "leaderboard",
          "immersive",
          "voice",
          "materials"
        ],
        "additionalProperties": false
      },
      "Bootstrap": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "modules": {
            "$ref": "#/components/schemas/ModuleFlags"
          },
          "serverNow": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "user",
          "modules",
          "serverNow"
        ],
        "additionalProperties": false
      },
      "ScenarioSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "scenario",
              "mega"
            ]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "serviceClass": {
            "type": "string"
          },
          "difficulty": {
            "type": "string"
          },
          "estimatedMinutes": {
            "type": "integer"
          },
          "competencies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "publishedVersion": {
            "type": "integer",
            "nullable": true
          },
          "draftRevision": {
            "type": "integer"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "presentation": {
            "type": "string",
            "enum": [
              "text",
              "immersive"
            ]
          }
        },
        "required": [
          "id",
          "kind",
          "title",
          "description",
          "serviceClass",
          "difficulty",
          "estimatedMinutes",
          "competencies",
          "publishedVersion",
          "draftRevision",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "ScenarioRecord": {
        "type": "object",
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/ScenarioSummary"
          },
          "definition": {
            "$ref": "#/components/schemas/ScenarioDefinition"
          },
          "revision": {
            "type": "integer"
          },
          "publishedVersion": {
            "type": "integer",
            "nullable": true
          }
        },
        "required": [
          "summary",
          "definition",
          "revision",
          "publishedVersion"
        ],
        "additionalProperties": false
      },
      "ValidationResult": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            }
          }
        },
        "required": [
          "valid",
          "issues"
        ],
        "additionalProperties": false
      },
...
      "SourceRef": {
        "type": "object",
        "properties": {
          "document": {
            "type": "string"
          },
          "section": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "document",
          "section"
        ],
        "additionalProperties": false
      },
      "ScenarioDefinition": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "integer",
            "enum": [
              1,
              2
            ]
          },
          "scene": {
            "$ref": "#/components/schemas/SceneDefinition"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "scenario",
              "mega"
            ]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "serviceClass": {
            "type": "string",
            "enum": [
              "standard",
              "comfort",
              "business",
              "first",
              "any"
            ]
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ]
          },
          "estimatedMinutes": {
            "type": "number"
          },
          "timerMode": {
            "type": "string",
            "enum": [
              "scenario",
              "step"
            ]
          },
          "competencies": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "communication",
                "service",
                "safety",
                "conflict"
              ]
            }
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            }
          },
          "startNodeId": {
            "type": "string"
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            }
          },
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            }
          },
          "childScenarioIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "schemaVersion",
          "id",
          "kind",
          "title",
          "description",
          "serviceClass",
          "difficulty",
          "estimatedMinutes",
          "competencies",
          "sources",
          "startNodeId",
          "nodes",
          "edges",
          "childScenarioIds"
        ],
        "additionalProperties": false
      },
      "ValidationIssue": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "nodeId": {
            "type": "string"
          },
          "edgeId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
...
      "SceneDefinition": {
        "type": "object",
        "properties": {
          "manifestVersion": {
            "type": "integer",
            "enum": [
              1
            ]
          },
          "brief": {
            "type": "string"
          },
          "objective": {
            "type": "string"
          },
          "rules": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "trainClass": {
            "type": "string",
            "enum": [
              "standard",
              "comfort",
              "business",
              "first"
            ]
          },
          "spawn": {
            "$ref": "#/components/schemas/WorldPoint"
          },
          "anchors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "label": {
                  "type": "string"
                },
                "x": {
                  "type": "number"
                },
                "z": {
                  "type": "number"
                },
                "radius": {
                  "type": "number"
                },
                "kind": {
                  "type": "string",
                  "enum": [
                    "passenger",
                    "radio",
                    "service",
                    "seat",
                    "exit"
                  ]
                }
              },
              "required": [
                "id",
                "label",
                "x",
                "z",
                "radius",
                "kind"
              ],
              "additionalProperties": false
            }
          },
          "passenger": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "age": {
                "type": "integer"
              },
              "description": {
                "type": "string"
              },
              "anchorId": {
                "type": "string"
              },
              "initialLine": {
                "type": "string"
              },
              "facing": {
                "type": "number",
                "minimum": 0,
                "maximum": 360
              }
            },
            "required": [
              "name",
              "age",
              "description",
              "anchorId",
              "initialLine"
            ],
            "additionalProperties": false
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "label": {
                  "type": "string"
                },
                "anchorId": {
                  "type": "string"
                },
                "prefab": {
                  "type": "string",
                  "enum": [
                    "blanket",
                    "cleaning_kit",
                    "bag",
                    "marker"
                  ]
                }
              },
              "required": [
                "id",
                "label",
                "anchorId",
                "prefab"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "manifestVersion",
          "trainClass",
          "spawn",
          "anchors",
          "passenger",
          "items"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Войти",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bootstrap"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "tags": [
          "Вход"
        ],
        "operationId": "post__auth_login"
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Зарегистрировать проводника и открыть сеанс",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  },
                  "lastName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 200
                  },
                  "password": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 128
                  }
                },
                "required": [
                  "firstName",
                  "lastName",
                  "email",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Создано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bootstrap"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "tags": [
          "Вход"
        ],
        "operationId": "post__auth_register"
      }
    },
...
    "/editor/scenarios": {
      "get": {
        "summary": "Черновики автора",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScenarioSummary"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "tags": [
          "Редактор"
        ],
        "operationId": "get__editor_scenarios"
      },
      "post": {
        "summary": "Создать черновик",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "kind": {
                    "type": "string",
                    "enum": [
                      "scenario",
                      "mega"
                    ]
                  }
                },
                "required": [
                  "title",
                  "kind"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Создано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioRecord"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "tags": [
          "Редактор"
        ],
        "operationId": "post__editor_scenarios"
      }
    },
    "/editor/scenarios/{id}": {
      "get": {
        "summary": "Получить черновик",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioRecord"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "tags": [
          "Редактор"
        ],
        "operationId": "get__editor_scenarios__id_"
      },
      "put": {
        "summary": "Сохранить черновик",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "definition": {
                    "$ref": "#/components/schemas/ScenarioDefinition"
                  },
                  "expectedRevision": {
                    "type": "integer"
                  }
                },
                "required": [
                  "definition",
                  "expectedRevision"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioRecord"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "tags": [
          "Редактор"
        ],
        "operationId": "put__editor_scenarios__id_"
      }
    },
    "/editor/scenarios/{id}/validate": {
      "post": {
        "summary": "Проверить граф",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "definition": {
                    "$ref": "#/components/schemas/ScenarioDefinition"
                  }
                },
                "required": [
                  "definition"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "tags": [
          "Редактор"
        ],
        "operationId": "post__editor_scenarios__id__validate"
      }
    },
    "/editor/scenarios/{id}/publish": {
      "post": {
        "summary": "Опубликовать неизменяемую версию",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "expectedRevision": {
                    "type": "integer"
                  }
                },
                "required": [
                  "expectedRevision"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioRecord"
                }
              }
            }
          },
          "422": {
            "description": "Некорректный граф",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "tags": [
          "Редактор"
        ],
        "operationId": "post__editor_scenarios__id__publish"
      }
    },
...
    "/openapi.json": {
      "get": {
        "summary": "Спецификация OpenAPI",
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "tags": [
          "Кабинет и система"
        ],
        "operationId": "get__openapi_json"
      }
    }
  }
}